# packages/api-types/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/check-username": {
      "get": {
        "description": "로그인/회원가입 전 해당 아이디의 가입 여부를 확인",
        "operationId": "AuthController_checkUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "description": "확인할 아이디",
            "schema": {
              "example": "chatnode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "확인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUsernameResponseDto"
                }
              }
            }
          }
        },
        "summary": "아이디 중복 확인",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sign-in": {
      "post": {
        "description": "아이디와 비밀번호로 로그인. 기존 활성 세션을 모두 차단 후 새 세션 생성. httpOnly 쿠키로 access_token, refresh_token 발급",
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공 (쿠키에 토큰 설정)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "아이디 또는 비밀번호 불일치"
          }
        },
        "summary": "로그인",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sign-up": {
      "post": {
        "description": "새 계정 생성 후 자동 로그인. httpOnly 쿠키로 access_token, refresh_token 발급",
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "회원가입 성공 (쿠키에 토큰 설정)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 사용 중인 아이디"
          }
        },
        "summary": "회원가입",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "refresh_token으로 새 access_token과 refresh_token을 발급. 기존 세션은 차단되고 새 세션 생성. 차단된 세션의 토큰으로 요청 시 탈취로 간주하여 해당 유저의 전체 세션 차단",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "토큰 갱신 성공 (쿠키 + body에 토큰 반환)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않거나 만료된 refresh_token / 탈취 감지",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          }
        },
        "summary": "토큰 갱신 (Refresh Token Rotation)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sign-out": {
      "delete": {
        "description": "현재 유저의 모든 활성 세션을 차단하고 인증 쿠키 삭제. 인증 필요",
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 요청"
          }
        },
        "summary": "로그아웃",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/force-logout/{targetUserId}": {
      "delete": {
        "description": "대상 유저의 모든 활성 세션을 차단. ADMIN 권한 필요",
        "operationId": "AuthController_forceLogout",
        "parameters": [
          {
            "name": "targetUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "강제 로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "ADMIN 권한 필요"
          }
        },
        "summary": "관리자 강제 로그아웃",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/channel": {
      "get": {
        "operationId": "ChannelController_getChannels",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "마지막으로 조회한 채널 ID (커서)",
            "schema": {
              "example": "e6f436e9-7f4f-4b52-95e3-ed43c07ad97b",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "조회할 항목 수",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetChannelsResponseDto"
                }
              }
            }
          }
        },
        "summary": "채널 목록 조회",
        "tags": [
          "channel"
        ]
      },
      "post": {
        "operationId": "ChannelController_createChannel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "채널 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateChannelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 채널 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          }
        },
        "summary": "채널 생성",
        "tags": [
          "channel"
        ]
      }
    },
    "/api/v1/channel/{channelId}/join": {
      "post": {
        "description": "채널에 입장합니다. 닉네임 미입력 시 아이디가 닉네임으로 설정됩니다. 비밀방인 경우 비밀번호 필수",
        "operationId": "ChannelController_joinChannel",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "입장 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinChannelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "비밀번호 미입력",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "비밀번호 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "채널을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 사용 중인 닉네임",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          }
        },
        "summary": "채널 입장",
        "tags": [
          "channel"
        ]
      }
    },
    "/api/v1/channel/dm": {
      "post": {
        "description": "대상 유저와 1:1 DM 채널을 생성합니다",
        "operationId": "ChannelController_createDm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DM 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDmResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "자기 자신과 DM 생성 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "대상 유저를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 해당 유저와의 DM이 존재",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullPayloadResponseDto"
                }
              }
            }
          }
        },
        "summary": "DM 생성",
        "tags": [
          "channel"
        ]
      }
    }
  },
  "info": {
    "title": "chat node api docs",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "에러 메시지"
          },
          "status": {
            "type": "number",
            "example": "401"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "CheckUsernamePayload": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean",
            "description": "아이디 존재 여부",
            "example": true
          }
        },
        "required": [
          "exists"
        ]
      },
      "CheckUsernameResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/CheckUsernamePayload"
          }
        },
        "required": [
          "payload"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "아이디",
            "example": "chatnode"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "password123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "NullPayloadResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "payload"
        ]
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "아이디 (5~20자)",
            "example": "chatnode",
            "minLength": 5,
            "maxLength": 20
          },
          "password": {
            "type": "string",
            "description": "비밀번호 (8자 이상)",
            "example": "password123",
            "minLength": 8
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AuthTokensPayload": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/AuthTokensPayload"
          }
        },
        "required": [
          "payload"
        ]
      },
...
      "LastMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "안녕하세요!"
          },
          "senderName": {
            "type": "string",
            "example": "사용자"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-25T12:00:00.000Z"
          }
        },
        "required": [
          "content",
          "senderName",
          "createdAt"
        ]
      },
      "ChannelItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "e6f436e9-7f4f-4b52-95e3-ed43c07ad97b"
          },
          "name": {
            "type": "string",
            "example": "테스트 채널"
          },
          "description": {
            "type": "string",
            "example": "테스트를 위한 채널입니다."
          },
          "profileImageUrl": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          },
          "lastMessage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LastMessageDto"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "example": 3
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-25T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "isPrivate",
          "unreadCount",
          "createdAt"
        ]
      },
      "GetChannelsPayload": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelItemDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "다음 페이지 커서 (null이면 마지막 페이지)",
            "example": "e6f436e9-7f4f-4b52-95e3-ed43c07ad97b"
          }
        },
        "required": [
          "channels"
        ]
      },
      "GetChannelsResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "example": null,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/GetChannelsPayload"
          }
        },
        "required": [
          "payload"
        ]
      },
      "JoinChannelDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "닉네임 (미입력 시 아이디로 설정)",
            "example": "사용자"
          },
          "password": {
            "type": "string",
            "description": "채널 비밀번호 (비밀방인 경우 필수)",
            "example": "123456"
          }
        }
      },
      "JoinChannelPayload": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string",
            "example": "e6f436e9-7f4f-4b52-95e3-ed43c07ad97b"
          },
          "role": {
            "type": "string",
            "example": "USER"
          },
          "displayName": {
            "type": "string",
            "example": "사용자"
          }
        },
        "required": [
          "channelId",
          "role",
          "displayName"
        ]
      },
      "JoinChannelResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "example": null,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/JoinChannelPayload"
          }
        },
        "required": [
          "payload"
        ]
      },
      "CreateChannelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "채널 이름 (2~50자)",
            "example": "테스트 채널",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "채널 설명 (2~200자)",
            "example": "테스트를 위한 채널입니다.",
            "minLength": 2,
            "maxLength": 200
          },
          "password": {
            "type": "string",
            "description": "비밀번호 (optional)",
            "example": "123456"
          },
          "profileImageUrl": {
            "type": "string",
            "description": "채널 프로필 이미지 (optional)",
            "example": "https://example.com/image.jpg"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "CreateChannelPayload": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string"
          }
        },
        "required": [
          "channelId"
        ]
      },
      "CreateChannelResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "example": null,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/CreateChannelPayload"
          }
        },
        "required": [
          "payload"
        ]
      },
      "CreateDmDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "description": "대화 상대 유저 ID",
            "example": "e6f436e9-7f4f-4b52-95e3-ed43c07ad97b"
          }
        },
        "required": [
          "targetUserId"
        ]
      },
      "CreateDmPayload": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string"
          }
        },
        "required": [
          "channelId"
        ]
      },
      "CreateDmResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true,
            "example": null,
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDto"
              }
            ]
          },
          "payload": {
            "$ref": "#/components/schemas/CreateDmPayload"
          }
        },
        "required": [
          "payload"
        ]
      }
    }
  },
...
}